refactor(routes): move confidence-control into its own findings-display module
