+++ companion/src/routes/aiSynthesis.ts
@@ -15,7 +15,6 @@ import {
   type NewHypothesis,
 } from "../analysis/hypothesis.js";
 import type { InvestigationQuestion, QuestionStatus } from "../analysis/stateTypes.js";
-import type { ConfidenceControl } from "../analysis/confidenceControl.js";
 import { STARRED_LABEL, type SuperQuery } from "../analysis/superTimeline.js";
 import { PresidioApprovalRequired } from "../analysis/presidio.js";
 import { sendPipelineError } from "./presidioApproval.js";
@@ -37,8 +36,6 @@ import { renderStandalonePresentation } from "../reports/presentationExport.js";
  *   - events/:eid/explain   — explain one forensic event in context (ephemeral).
  *   - questions             — pin an analyst key question so synthesis preserves + answers it.
  *   - hypotheses (CRUD)     — status-tracked investigative hypotheses.
- *   - confidence-control (GET/PUT) — the findings min-confidence DISPLAY floor (#226); left here by
- *     the findings domain because it belongs to AI output presentation, not analyst annotations.
  *   - narrative (POST generate / PUT save-edit) — prose narrative timeline.
  *   - executive-summary / remediation-plan — one-shot management/defensive AI generations.
  *   - adversary-hints (GET) + adversary-hints/hunt-technique (POST) — offline ATT&CK-group hints +
@@ -760,77 +757,6 @@ export function registerAiSynthesisRoutes(app: Express, ctx: RouteContext): void
       if (!removed) return res.status(404).json({ error: "hypothesis not found" });
       options.onHypotheses?.(req.params.id);
       return res.status(204).end();
-    } catch (err) {
-      return sendPipelineError(res, err);
-    }
-  });
-
-  // Findings-panel display preferences: the min-confidence floor (#226) and the two finding-origin
-  // lenses. Per-case and persisted so they survive a page reload; purely display (nothing is
-  // removed from state). `minConfidence: null` means "show all" (0); each lens defaults to false.
-  const confidenceControlBody = (c: ConfidenceControl) => ({
-    minConfidence: c.minConfidence ?? null,
-    hideAutoFindings: c.hideAutoFindings ?? false,
-    hideGapFindings: c.hideGapFindings ?? false,
-  });
-
-  app.get("/cases/:id/confidence-control", async (req: Request, res: Response) => {
-    if (!options.confidenceControlStore)
-      return res.status(501).json({ error: "confidence control not configured" });
-    try {
-      return res.status(200).json(confidenceControlBody(await options.confidenceControlStore.load(req.params.id)));
-    } catch (err) {
-      return sendPipelineError(res, err);
-    }
-  });
-
-  app.put("/cases/:id/confidence-control", async (req: Request, res: Response) => {
-    if (!options.confidenceControlStore)
-      return res.status(501).json({ error: "confidence control not configured" });
-    const body = (req.body ?? {}) as Record<string, unknown>;
-    // PATCH ONLY THE KEYS THE REQUEST CARRIES. The dashboard writes the floor (debounced, on
-    // keystrokes) and the lenses (immediate, on click) from two independent paths, so a PUT naming
-    // one must not disturb the others. Reading an ABSENT minConfidence as "clear it" — which is
-    // what this handler used to do — would let a checkbox click wipe the analyst's floor.
-    const patch: ConfidenceControl = {};
-    if ("minConfidence" in body) {
-      const raw = body.minConfidence;
-      const cleared = raw === null || raw === undefined || raw === "";
-      if (!cleared && (typeof raw !== "number" || !Number.isFinite(raw) || raw < 0 || raw > 100)) {
-        return res.status(400).json({ error: "minConfidence must be a number 0-100, or null" });
-      }
-      patch.minConfidence = cleared ? undefined : raw;
-    }
-    for (const key of ["hideAutoFindings", "hideGapFindings"] as const) {
-      if (!(key in body)) continue;
-      const raw = body[key];
-      if (raw !== null && raw !== undefined && typeof raw !== "boolean") {
-        return res.status(400).json({ error: `${key} must be a boolean, or null` });
-      }
-      patch[key] = raw == null ? undefined : raw;
-    }
-    try {
-      await options.confidenceControlStore.set(req.params.id, patch);
-      options.onConfidenceControl?.(req.params.id);
-      const saved = await options.confidenceControlStore.load(req.params.id);
-      const detail =
-        [
-          "minConfidence" in patch
-            ? saved.minConfidence === undefined
-              ? "minConfidence cleared"
-              : `minConfidence set to ${saved.minConfidence}`
-            : null,
-          "hideAutoFindings" in patch ? `hideAutoFindings ${saved.hideAutoFindings ?? false}` : null,
-          "hideGapFindings" in patch ? `hideGapFindings ${saved.hideGapFindings ?? false}` : null,
-        ]
-          .filter(Boolean)
-          .join("; ") || "no change";
-      void logActivity(options.activityLogStore, options.onActivity, req.params.id, {
-        category: "settings",
-        action: "confidence-control",
-        detail,
-      });
-      return res.status(200).json(confidenceControlBody(saved));
     } catch (err) {
       return sendPipelineError(res, err);
     }
