+++ companion/tests/architecture/route-inventory.json
@@ -332,8 +332,6 @@
     "ROUTE POST /cases/:id/hypotheses",
     "ROUTE PATCH /cases/:id/hypotheses/:hid",
     "ROUTE DELETE /cases/:id/hypotheses/:hid",
-    "ROUTE GET /cases/:id/confidence-control",
-    "ROUTE PUT /cases/:id/confidence-control",
     "ROUTE GET /cases/:id/adversary-hints",
     "ROUTE POST /cases/:id/adversary-hints/hunt-technique",
     "ROUTE POST /cases/:id/memory/next-steps",
@@ -341,6 +339,8 @@
     "ROUTE GET /cases/:id/presentation",
     "ROUTE GET /cases/:id/present/export",
     "ROUTE GET /cases/:id/present",
+    "ROUTE GET /cases/:id/confidence-control",
+    "ROUTE PUT /cases/:id/confidence-control",
     "ROUTE GET /cases/:id/deep-pass/preview",
     "ROUTE POST /cases/:id/deep-pass",
     "ROUTE GET /cases/:id/report/interactive",
